main/java/service/ClienteService.java
package main.java.service;

import main.java.DAO.ClienteDAO;
import main.java.domain.Cliente;
import main.java.DAO.IClienteDAO;
import main.java.exception.TipoChaveNaoEncontradaException;

public class ClienteService implements IClienteService {

    private IClienteDAO clienteDAO;

    public ClienteService(IClienteDAO clienteDAO){
        this.clienteDAO = clienteDAO;
    }

    @Override
    public Boolean salvar(Cliente cliente) throws TipoChaveNaoEncontradaException {
        return clienteDAO.cadastrar(cliente);
    }

    @Override
    public Cliente buscarPorCPF(Long cpf) {
        return clienteDAO.consultar(cpf);
    }

    @Override
    public void excluir(Long cpf) {

    }

    @Override
    public void alterar(Cliente cliente) throws TipoChaveNaoEncontradaException{
        clienteDAO.alterar(cliente);

    }
}
